+++ CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.2](https://github.com/basaltinc/bedrock/compare/v0.17.1...v0.17.2) (2018-12-08)
+
+
+### Bug Fixes
+
+* **bedrock:** improve css and js config paths ([#146](https://github.com/basaltinc/bedrock/issues/146)) ([9539a54](https://github.com/basaltinc/bedrock/commit/9539a54)), closes [#145](https://github.com/basaltinc/bedrock/issues/145)
+* **bedrock:** type fix on config processing ([f8ad374](https://github.com/basaltinc/bedrock/commit/f8ad374))
+
+
+### Features
+
+* setup releases ([fdf7fca](https://github.com/basaltinc/bedrock/commit/fdf7fca)), closes [#145](https://github.com/basaltinc/bedrock/issues/145)
+
+
+
+
+
+## [0.17.1](https://github.com/basaltinc/bedrock/compare/v0.17.0...v0.17.1) (2018-12-08)
+
+
+### Bug Fixes
+
+* **bedrock:** add proper error page for bad pattern urls ([#139](https://github.com/basaltinc/bedrock/issues/139)) ([b992a4f](https://github.com/basaltinc/bedrock/commit/b992a4f)), closes [#137](https://github.com/basaltinc/bedrock/issues/137)
+* **bedrock:** fix warning about sidebar collapse ([56316cc](https://github.com/basaltinc/bedrock/commit/56316cc)), closes [#107](https://github.com/basaltinc/bedrock/issues/107)
+
+
+
+
+
 # [0.17.0](https://github.com/basaltinc/bedrock/compare/v0.16.6...v0.17.0) (2018-12-08)
 
 **Note:** Version bump only for package @basalt/bedrock-monorepo